+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+<a name="0.2.0"></a>
+### 0.2.0
+#### :rocket: Enhancements
+* Support running releases of Liferay Portal
+
 <a name="0.1.2"></a>
 ### 0.1.2
 #### :rocket: Enhancements